requirements.txt
google-api-python-client
streamlit
streamlit-cookies-manager
pandas
google-generativeai
bcrypt
fpdf
numpy
PyMuPDF
groq


